requirements/requirements.txt
coreapi==2.3.3
Django==3.1.5
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
